fix: avoid crashing on web
